[DP][DDCE-1282] Update to sbt 1.13.13 and scala 2.12.12
