--- .travis.yml
@@ -1,14 +1,0 @@
-
-sudo: false
-language: scala
-scala:
-- 2.11.6
-jdk:
-- oraclejdk8
-cache:
-  directories:
-    - '$HOME/.ivy2/cache'
-branches:
-  except:
-    - master
-dist: trusty
