+++ build.sbt
@@ -1,0 +1,56 @@
+import uk.gov.hmrc._
+import DefaultBuildSettings._
+import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin.publishingSettings
+import play.sbt.PlayImport.PlayKeys
+import play.sbt.routes.RoutesKeys.{InjectedRoutesGenerator, routesGenerator}
+import sbt.Keys._
+import sbt._
+import scoverage.ScoverageKeys
+import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin
+import TestPhases._
+import uk.gov.hmrc.versioning.SbtGitVersioning
+import uk.gov.hmrc.versioning.SbtGitVersioning.autoImport.majorVersion
+
+
+val appName: String = "ers-file-validator"
+
+lazy val appDependencies: Seq[ModuleID] = AppDependencies.apply()
+lazy val plugins: Seq[Plugins] = Seq(play.sbt.PlayScala, SbtAutoBuildPlugin, SbtGitVersioning, SbtDistributablesPlugin, SbtArtifactory)
+lazy val testPhases = TestPhases
+
+lazy val scoverageSettings = {
+  Seq(
+    ScoverageKeys.coverageExcludedPackages := "<empty>;app.*;config.*;testOnlyDoNotUseInAppConf.*;views.*;uk.gov.hmrc.*;prod.*;models.*;services.ERSRequest",
+    ScoverageKeys.coverageMinimum := 87,
+    ScoverageKeys.coverageFailOnMinimum := false,
+    ScoverageKeys.coverageHighlighting := true,
+    parallelExecution in Test := false
+  )
+}
+
+lazy val microservice = Project(appName, file("."))
+  .enablePlugins(plugins: _*)
+  .settings(scoverageSettings: _*)
+  .settings(publishingSettings)
+  .settings(scalaSettings: _*)
+  .settings(defaultSettings(): _*)
+  .settings(
+    targetJvm := "jvm-1.8",
+    scalaVersion := "2.12.12",
+    libraryDependencies ++= appDependencies,
+    parallelExecution in Test := false,
+    fork in Test := false,
+    retrieveManaged := true,
+    routesGenerator := InjectedRoutesGenerator
+  )
+  .settings(inConfig(TemplateTest)(Defaults.testSettings): _*)
+  .configs(IntegrationTest)
+  .settings(inConfig(TemplateItTest)(Defaults.itSettings): _*)
+  .settings(
+    Keys.fork in IntegrationTest := false,
+    unmanagedSourceDirectories in IntegrationTest := (baseDirectory in IntegrationTest) (base => Seq(base / "it")).value,
+    addTestReportOption(IntegrationTest, "int-test-reports"),
+    parallelExecution in IntegrationTest := false)
+  .settings(resolvers += Resolver.bintrayRepo("hmrc", "releases"), resolvers += Resolver.jcenterRepo)
+  .settings(majorVersion := 1)
+  .settings(PlayKeys.playDefaultPort := 9226)
