--- project/MicroService.scala
@@ -1,97 +1,0 @@
-
-import play.sbt.PlayImport.PlayKeys
-import play.sbt.routes.RoutesKeys.{InjectedRoutesGenerator, routesGenerator}
-import sbt.Keys._
-import sbt.Tests.{Group, SubProcess}
-import sbt._
-import scoverage.ScoverageKeys
-import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin
-import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin._
-import uk.gov.hmrc.versioning.SbtGitVersioning
-import uk.gov.hmrc.versioning.SbtGitVersioning.autoImport.majorVersion
-
-trait MicroService {
-
-  import uk.gov.hmrc._
-  import DefaultBuildSettings._
-  import TestPhases._
-
-
-  val appName: String
-
-  lazy val appDependencies: Seq[ModuleID] = ???
-  lazy val plugins: Seq[Plugins] = Seq(play.sbt.PlayScala, SbtAutoBuildPlugin, SbtGitVersioning, SbtDistributablesPlugin, SbtArtifactory)
-  lazy val playSettings: Seq[Setting[_]] = Seq.empty
-
-  lazy val scoverageSettings = {
-    Seq(
-      ScoverageKeys.coverageExcludedPackages := "<empty>;app.*;config.*;testOnlyDoNotUseInAppConf.*;views.*;uk.gov.hmrc.*;prod.*;models.*;services.ERSRequest",
-      ScoverageKeys.coverageMinimum := 87,
-      ScoverageKeys.coverageFailOnMinimum := false,
-      ScoverageKeys.coverageHighlighting := true,
-      parallelExecution in Test := false
-    )
-  }
-
-  lazy val microservice = Project(appName, file("."))
-    .enablePlugins(plugins: _*)
-    .enablePlugins(SbtAutoBuildPlugin, SbtGitVersioning)
-    .enablePlugins(SbtDistributablesPlugin)
-    .settings(playSettings: _*)
-    .settings(scoverageSettings: _*)
-    .settings(publishingSettings: _*)
-    .settings(scalaSettings: _*)
-    .settings(defaultSettings(): _*)
-    .settings(
-      targetJvm := "jvm-1.8",
-      scalaVersion := "2.11.8",
-      libraryDependencies ++= appDependencies,
-      parallelExecution in Test := false,
-      fork in Test := false,
-      retrieveManaged := true,
-      routesGenerator := InjectedRoutesGenerator
-    )
-    .settings(Repositories.playPublishingSettings: _*)
-    .settings(inConfig(TemplateTest)(Defaults.testSettings): _*)
-    .configs(IntegrationTest)
-    .settings(inConfig(TemplateItTest)(Defaults.itSettings): _*)
-    .settings(
-      Keys.fork in IntegrationTest := false,
-      unmanagedSourceDirectories in IntegrationTest <<= (baseDirectory in IntegrationTest) (base => Seq(base / "it")),
-      addTestReportOption(IntegrationTest, "int-test-reports"),
-      testGrouping in IntegrationTest := TestPhases.oneForkedJvmPerTest((definedTests in IntegrationTest).value),
-      parallelExecution in IntegrationTest := false)
-    .settings(resolvers += Resolver.bintrayRepo("hmrc", "releases"), resolvers += Resolver.jcenterRepo)
-    .settings(majorVersion := 1)
-    .settings(PlayKeys.playDefaultPort := 9226)
-}
-
-private object TestPhases {
-
-  val allPhases = "tt->test;test->test;test->compile;compile->compile"
-  val allItPhases = "tit->it;it->it;it->compile;compile->compile"
-
-  lazy val TemplateTest = config("tt") extend Test
-  lazy val TemplateItTest = config("tit") extend IntegrationTest
-
-  def oneForkedJvmPerTest(tests: Seq[TestDefinition]) =
-    tests map {
-      test => new Group(test.name, Seq(test), SubProcess(ForkOptions(runJVMOptions = Seq("-Dtest.name=" + test.name))))
-    }
-}
-
-private object Repositories {
-
-  import uk.gov.hmrc._
-  import PublishingSettings._
-
-  lazy val playPublishingSettings: Seq[sbt.Setting[_]] = sbtrelease.ReleasePlugin.releaseSettings ++ Seq(
-
-    credentials += SbtCredentials,
-
-    publishArtifact in(Compile, packageDoc) := false,
-    publishArtifact in(Compile, packageSrc) := false
-  ) ++
-    publishAllArtefacts
-
-}
